Add text_parser and word_indexer Functions
This commit adds a text_parser function which takes a file and
returns a list of strings that were delimited by non alphanumeric
characters. This commit also adds the word_indexer function which
takes a list of strings and indexes them into a master index of
all strings that have been seen up to that point.